Chore: remove & clean & rename Files (working)
