--- README.md
@@ -1,5 +1,0 @@
-# 42_webserv
-
-개인 노션 https://www.notion.so/webserv-735a01bd55044703a92f82bbe4b92d3f
-
-팀 노션 https://www.notion.so/lightningcy/webserv-b755140bf4b941e3be42ac733f9eb7e3
